test_smf returns msglen for large message
This is the first part of large message optimization for compression.
This change also makes large non-compressed smf message decode faster because tcp will not call smf dissection if it does not have a big enough pdu.

Why does the code done like this to begin with?
This is partially my fault. Originally, the code to test smf has two functions, one to just test the header, the other to test AND compute the length. I saw a lot of common code in there and refactor it, but made the above error.