SMF changes to attachment subdissectors (#13)
* Pull from main (#12)

* Added Wireshark 4.0.5 submodule

* Added Getting Started Section to README

* Update installation instructions

* Update the README (#8)

* Update the README to make the major.minor version convention more clear.

* Minor edits to the README

* test_smf returns msglen for large message (#6)

This is the first part of large message optimization for compression.
This change also makes large non-compressed smf message decode faster because tcp will not call smf dissection if it does not have a big enough pdu.

Why does the code done like this to begin with?
This is partially my fault. Originally, the code to test smf has two functions, one to just test the header, the other to test AND compute the length. I saw a lot of common code in there and refactor it, but made the above error.

* Added @ main to each build job for build_plugin.yml (#10)

* Stronger smf type detection to detect smf header (#4)

* Stronger smf type detection to detect smf header

When smf-header is incorrectly detected, obsoleting some obsolete smf types may help remove some of these false-positives.

* Update based on comments

* Update after conversation with Ragnar

* Wireshark submodule for version 4.0.5 (#11)

Method to upgrade to 4.0.5.
I am using Windows PC.
Go into the wireshark subdirectory and execute the command:
git reset --hard v4.0.5

---------

Co-authored-by: Neel Patel <156137236+NeelPatel-Solace@users.noreply.github.com>
Co-authored-by: Ragnar <36005865+RagnarPaulson@users.noreply.github.com>

* SMF change for attachment subdissectors

Although the changes has solcache subdissector as a usecase, the approach is more generic for all subdissectors of smf attachments.
The change involves:
1) SMF subdissectors is attempted before the decoding based on attachment type.
2) The smf_subdissection_uat is loaded at dissection time rather than at smf dissector initialization. The reason is that the C dissectors are initialized at a time when the lua plugin dissectors are not yet loaded.

---------

Co-authored-by: Neel Patel <156137236+NeelPatel-Solace@users.noreply.github.com>
Co-authored-by: Ragnar <36005865+RagnarPaulson@users.noreply.github.com>