Binary Meta data function does not handle snaplen packet capture well
In Support we do a lot of snaplength packet captures. If a message contains binary metadata and only part of the data is captured in the binary metadata, the function call proto_tree_add_string() in dissect_bm() will fail which is caught by the proto dissector outside smf, which is fine.
However, the Info column is all messed up because the smf dissector function deletes the info column and the replaced column information is added after a successful return of the binary metadata dissector.

To avoid this, we could either avoid calling the binary metadata dissector altogether if the binary metadata is not fully available, or protect the proto_tree_add_string() call.

I have chosen a simpler approach by not calling the function if the data is not all there.