Correct Message Attachment variable usage
The variable param_info::attachment_length was used as both binary attachment length as well as full attachment length causing the confusion.
I also find that the variable param_info::xml_payload_length incorrectly used for convenience (it is set as the payload length for non-TRMSG).
In this PR, we introducing a new binary_payload_length variable to separate out from attachment_length and refactoring base on this introduction. The usage of xml_payload_length is also cleaned up.